+++ src/melee/ft/chara/ftNess/ftNs_SpecialLw.c
@@ -50,7 +50,7 @@ void ftNs_SpecialLwStart_Enter(HSD_GObj* gobj) // Ness's grounded PSI Magnet
         (s32) ness_attr->x74_PSI_MAGNET_RELEASE_LAG;
     temp_fp->mv.ns.speciallw.isRelease = 0;
     temp_fp->mv.ns.speciallw.gravityDelay =
-        (s32) ness_attr->x84_PSI_MAGNET_FRAMES_BEFORE_GRAVITY;
+        ness_attr->x84_PSI_MAGNET_FRAMES_BEFORE_GRAVITY;
     temp_fp->mv.ns.speciallw.x10 = 0;
     Fighter_ChangeMotionState(gobj, ftNs_MS_SpecialLwStart, 0, 0, 1, 0, NULL);
     ftAnim_8006EBA4(gobj);
@@ -72,7 +72,7 @@ void ftNs_SpecialAirLwStart_Enter(
         (s32) ness_attr->x74_PSI_MAGNET_RELEASE_LAG;
     temp_fp->mv.ns.speciallw.isRelease = 0;
     temp_fp->mv.ns.speciallw.gravityDelay =
-        (s32) ness_attr->x84_PSI_MAGNET_FRAMES_BEFORE_GRAVITY;
+        ness_attr->x84_PSI_MAGNET_FRAMES_BEFORE_GRAVITY;
     temp_fp->mv.ns.speciallw.x10 = 0;
     temp_fp->self_vel.y = 0.0f;
     temp_fp->self_vel.x /= ness_attr->x88_PSI_MAGNET_MOMENTUM_PRESERVATION;
@@ -250,7 +250,7 @@ void ftNs_SpecialLwHold_Anim(
     if (fp->mv.ns.speciallw.releaseLag > 0) {
         fp->mv.ns.speciallw.releaseLag--;
     }
-    if (((s32) fp->mv.ns.speciallw.releaseLag <= 0) &&
+    if ((fp->mv.ns.speciallw.releaseLag <= 0) &&
         ((s32) fp->mv.ns.speciallw.isRelease != 0))
     {
         if ((s32) fp->ground_or_air == GA_Ground) {
@@ -263,7 +263,7 @@ void ftNs_SpecialLwHold_Anim(
 
     temp_r31->mv.ns.speciallw.x10--;
 
-    if ((s32) temp_r31->mv.ns.speciallw.x10 <= 0) {
+    if (temp_r31->mv.ns.speciallw.x10 <= 0) {
         ft_80088478(temp_r31, 210081, 127, 64);
         temp_r31->mv.ns.speciallw.x10 = 40;
     }
@@ -288,7 +288,7 @@ void ftNs_SpecialAirLwHold_Anim(
         fp->mv.ns.speciallw.releaseLag =
             (s32) (fp->mv.ns.speciallw.releaseLag - 1);
     }
-    if (((s32) fp->mv.ns.speciallw.releaseLag <= 0) &&
+    if ((fp->mv.ns.speciallw.releaseLag <= 0) &&
         ((s32) fp->mv.ns.speciallw.isRelease != 0))
     {
         if ((s32) fp->ground_or_air == GA_Ground) {
@@ -301,7 +301,7 @@ void ftNs_SpecialAirLwHold_Anim(
 
     temp_r31->mv.ns.speciallw.x10 = (s32) (temp_r31->mv.ns.speciallw.x10 - 1);
 
-    if ((s32) temp_r31->mv.ns.speciallw.x10 <= 0) {
+    if (temp_r31->mv.ns.speciallw.x10 <= 0) {
         ft_80088478(temp_r31, 210081, 127, 64);
         temp_r31->mv.ns.speciallw.x10 = 40;
     }
@@ -458,7 +458,7 @@ static inline void getAttrStuff(HSD_GObj* arg0)
     ftNessAttributes* temp_r31 = temp_r30->dat_attrs;
     temp_r30->mv.ns.speciallw.turnFrames += -1;
 
-    if (((u32) temp_r30->cmd_vars[0] == 0U) &&
+    if ((temp_r30->cmd_vars[0] == 0U) &&
         ((returnStateVar(temp_r30->mv.ns.speciallw.turnFrames)) <=
          temp_r31->x78_PSI_MAGNET_UNK1))
     {
@@ -667,7 +667,7 @@ static inline void MagnetStateVarCalc(HSD_GObj* gobj)
 {
     Fighter* fp = GET_FIGHTER(gobj);
     fp->mv.ns.speciallw.x10 = fp->mv.ns.speciallw.x10 - 1;
-    if ((s32) fp->mv.ns.speciallw.x10 <= 0) {
+    if (fp->mv.ns.speciallw.x10 <= 0) {
         ft_80088478(fp, 210081, 127, 64);
         fp->mv.ns.speciallw.x10 = 40;
     }
@@ -701,7 +701,7 @@ void ftNs_SpecialLwHit_Anim(
     }
     if (!ftAnim_IsFramesRemaining(arg0)) {
         temp_r3_2 = arg0->user_data;
-        if (((s32) temp_r3_2->mv.ns.speciallw.releaseLag <= 0) &&
+        if ((temp_r3_2->mv.ns.speciallw.releaseLag <= 0) &&
             ((s32) temp_r3_2->mv.ns.speciallw.isRelease != 0))
         {
             if ((s32) temp_r3_2->ground_or_air == GA_Ground) {
@@ -922,7 +922,7 @@ void ftNs_AbsorbThink_DecideAction(
         pl_80040B8C(fp->player_id, fp->x221F_b4, heal_amount);
     }
 
-    fp->facing_dir = (float) fp->AbsorbAttr.x1A40_absorbHitDirection;
+    fp->facing_dir = fp->AbsorbAttr.x1A40_absorbHitDirection;
     msid = fp->motion_id;
 
     if ((msid != ftNs_MS_SpecialLwHit && msid != ftNs_MS_SpecialAirLwHit) ||
